getUserFromToken utility function
